remove dslash helper functions from interface to avoid cpp code there
